refactor: extract token-collector and git-metrics-collector from usage-manager
Reduce coupling in usage-manager.js by extracting token collection logic
into token-collector.js and git metrics logic into git-metrics-collector.js.
The public API (init, getMetrics) remains unchanged.

Closes #195
